i'm sorry for my self again

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,20 +11,20 @@
  * Return:the number of nodes
 */
 
-int print_list(const list_t *h)
+size_t print_list(const list_t *h)
 {
-int count = 0;
+size_t count = 0;
 
 
 while (h != NULL)
 {
 if (h->str == NULL)
-			printf("[0] (nil)\n");
-		else
-			printf("[%d] %s\n", h->len, h->str);
+printf("[0] (nil)\n");
+else
+printf("[%d] %s\n", h->len, h->str);
 
-		h = h->next;
-		count++;
+h = h->next;
+count++;
 
 }
